clarified cfp submission steps and moved AddSpeakerToTalkForm into talks.py file
