+++ templates/conference/cfp/step2_add_speaker.html
@@ -10,7 +10,7 @@
 
     <div class="row">
         <div class="col-md-12">
-            <h1>Add Speaker Information</h1>
+            <h1>Your Speaker Profile</h1>
         </div>
     </div>
 
@@ -20,7 +20,7 @@
                 {% csrf_token %}
                 {{ speaker_form|crispy }}
 
-                <button class="btn btn-success btn-lg btn-block">Add Speaker</button>
+                <button class="btn btn-success btn-lg btn-block">Submit Proposal</button>
             </form>
         </div>
